+++ src/migrator/Migrator.ts
@@ -12,11 +12,20 @@ import { MigrateError } from "../errors";
 
 import { MigrateConfig } from "../types/migrations";
 
+import { Linker } from "../deployer/Linker";
 import { Deployer } from "../deployer/Deployer";
 import { Verifier } from "../verifier/Verifier";
 
 import { Stats } from "../tools/Stats";
-import { Reporter } from "../tools/reporters/Reporter";
+
+import { initReporter, reporter } from "../tools/reporters/Reporter";
+import { transactionRunner } from "../tools/runners/TransactionRunner";
+
+import { initNetworkManager } from "../tools/network/NetworkManager";
+
+import { TransactionProcessor } from "../tools/storage/TransactionProcessor";
+import { MigrateStorage } from "../tools/storage/MigrateStorage";
+import { ArtifactProcessor } from "../tools/storage/ArtifactProcessor";
 
 export class Migrator {
   private readonly _deployer: Deployer;
@@ -96,5 +105,19 @@ export class Migrator {
 
   private _getMigrationNumber(file: string) {
     return parseInt(basename(file));
+  }
+
+  public static async initializeDependencies(hre: HardhatRuntimeEnvironment): Promise<void> {
+    Linker.setConfig(hre.config.migrate);
+    TransactionProcessor.setConfig(hre.config.migrate);
+
+    initNetworkManager(hre);
+    await initReporter(hre.config.migrate);
+
+    if (!hre.config.migrate.continue) {
+      MigrateStorage.clearAll();
+    }
+
+    await ArtifactProcessor.parseArtifacts(hre);
   }
 }